satisfacao_app/requirements.txt
Flask
flask
psycopg2-binary
gunicorn